Silence some warnings when compiling taskmgr with MSVC2005
svn path=/trunk/; revision=30306

+++ reactos/base/applications/taskmgr/perfpage.c
@@ -441,8 +441,8 @@ DWORD WINAPI PerformancePageRefreshThread(void *lpParameter)
             /*
              *  Get the memory usage
              */
-            CommitChargeTotal = (ULONGLONG)PerfDataGetCommitChargeTotalK();
-            CommitChargeLimit = (ULONGLONG)PerfDataGetCommitChargeLimitK();
+            CommitChargeTotal = PerfDataGetCommitChargeTotalK();
+            CommitChargeLimit = PerfDataGetCommitChargeLimitK();
             nBarsUsed1 = CommitChargeLimit ? ((CommitChargeTotal * 100) / CommitChargeLimit) : 0;
 
             PhysicalMemoryTotal = PerfDataGetPhysicalMemoryTotalK();
